% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_utils.R
\name{check_arg_names}
\alias{check_arg_names}
\title{Check the names of the passed input arguments and trigger error if argument
is not supported by any of the run functions.}
\usage{
check_arg_names(dot_args, version)
}
\arguments{
\item{dot_args}{Character vector of provided input arguments}

\item{version}{Version of the SWAT project, one of 'plus', '2012'}
}
\value{
Error message if argument is not supported
}
\description{
Check the names of the passed input arguments and trigger error if argument
is not supported by any of the run functions.
}
\keyword{internal}
