% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{compare_tables}
\alias{compare_tables}
\title{Compare a table in the current SWAT simulation with the one saved in 'save_file'}
\usage{
compare_tables(tbl_sim, tbl_df, txt, convert = FALSE)
}
\arguments{
\item{tbl_sim}{Table defined for the current simulation}

\item{tbl_df}{Corresponding table saved in the data bases}

\item{txt}{Text string to be added in the error message.}

\item{convert}{Whether to compare the tables after conversion to matrices.}
}
\description{
Compare a table in the current SWAT simulation with the one saved in 'save_file'
}
\keyword{internal}
