% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_output.R
\name{define_output}
\alias{define_output}
\title{Define SWAT output variables}
\usage{
define_output(file, variable = NULL, unit = NULL, label = NULL)
}
\arguments{
\item{file}{The SWAT output file to read.

For SWAT+ simulations with \code{run_swatplus()} all object files can be used
which are defined in the input file 'print.prt'. All file definitions
require to add the time interval extension in which the outputs should be
returned. The options for the time interval are \code{'day'} for daily , \code{'mon'}
for monthly, \code{'yr'} for yearly, and \code{'aa'} for average annual outputs.
Additionally, the outputs \code{'fdcout'} for flow duration curve outputs,
\code{'mgtout'} for management outputs, or \verb{'basin_crop_yld_<yr or aa>} for
basin crop yields can be defined.

For SWAT2012 simulations with \code{run_swat2012()} valid inputs are
\code{'rch'}, \code{'sub'}, \code{'hru'}, and \code{'sed'} for the
respective SWAT2012 output files 'output.rch', 'output.sub', output.hru',
and 'output.sed'.}

\item{variable}{Output variable that is extracted from the defined SWAT
output \code{file}.

A \code{variable} must be defined with the name which the variable has in the
respective output table. The variable names are case sensitive (SWAT2012
variables e.g. are usually all caps). In the \code{variable} definition units
(which are part of the names in SWAT2012 output files) should be excluded
from the name. Spaces in variable names must be replaced by an underscore
(e.g. P TOT is defined as \code{'P_TOT'}).

Some SWAT+ output files have specific output variable names. For \code{'mgtout'}
outputs the variables \code{'yld'}, \code{'phu'}, \code{'bioms'}, \code{'wat_strs'},
,\code{'aer_strs'},\code{'tmp_strs'}, \code{'n_strs'}, and \code{'p_strs'} can be returned. For
\verb{'basin_crop_yld_*} outputs the variables \code{'harv_area'}, \code{'yld_total'}, and
\code{'yld'} can be returned.}

\item{unit}{The spatial unit for which outputs are returned.

In SWAT+ output files spatial units are in most cases defined with the
'unit' column and are e.g. the HRU IDs in \code{'hru_*'} output files, or
channel IDs e.g. in \code{'channel_sd'} or \verb{'channel_sdmorph_*} output files
(same applies to all other object types).

In SWAT2012 output files \code{unit} is the  the reach, subbasin, or HRU ID
defined by the columns 'RCH', 'SUB', or 'HRU' in the respective SWAT2012
output file.}

\item{label}{Optional crop or output labels to retain from labeled outputs.}
}
\value{
A table with the columns 'file', 'variable', 'unit', and 'label'.
}
\description{
Define the SWAT+ or SWAT2012 output variables which are returned for a SWAT
model run with \code{run_swatplus()} or \code{run_swat2012()}. The defined outputs are
passed with the input argument \code{output} in either one of the run functions.
}
\examples{
# A single variable can be defined as follows (e.g. "FLOW_OUT" for
# the reaches 1 and 5):

out_flow <- define_output(file = "rch",
                          variable = "FLOW_OUT",
                          unit = c(1,5))
# In this case the the variable name of the returned output is then
# the same as defined with 'variable', here "FLOW_OUT"

# If a custom variable name is preferred for the returned output,
# the output must be defined as named list:

out_flow <- list(discharge = define_output(file = "rch",
                                           variable = "FLOW_OUT",
                                           unit = c(1,5)))


# Define the discharge for the RCH units 1 to 5 and the
# nitrate-nitrogen load and ET for the unit 5:

out_def <- list(flow = define_output(file = "rch",
                                     variable = "FLOW_OUT",
                                     unit = 1:5),
                no3  = define_output(file = "rch",
                                     variable = "NO3_OUT",
                                     unit = 5),
                et_a = define_output(file = "sub",
                                     variable = "ET",
                                     unit = 5))

}
