% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_handle_output.R
\name{get_fwf_positions}
\alias{get_fwf_positions}
\title{Derive the fixed with positions of the columns in the SWAT output files}
\usage{
get_fwf_positions(output_i, thread_path, tbl_pos)
}
\arguments{
\item{output_i}{Output defined to read from the SWAT model results}

\item{thread_path}{Path to respective thread where SWAT was executed}

\item{tbl_pos}{Position of the first data row in the output file.}
}
\description{
Derive the fixed with positions of the columns in the SWAT output files
}
\keyword{internal}
