% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{group_values}
\alias{group_values}
\title{Convert the information on available runs for the simulated variables into
strings that are printed}
\usage{
group_values(vals, sep = ":")
}
\arguments{
\item{vals}{overview table that provides meta data for all simulation runs for
all variables saved in the data bases}

\item{sep}{Separator used between the bounds of a range.}
}
\description{
Convert the information on available runs for the simulated variables into
strings that are printed
}
\keyword{internal}
