% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_outputs.R
\name{initialize_run_info}
\alias{initialize_run_info}
\title{Prepare run info for the simulation experiment}
\usage{
initialize_run_info(model_setup, output, project_path, run_path)
}
\arguments{
\item{model_setup}{List of model configurations}

\item{output}{Table of defined output variables}

\item{project_path}{Path to the SWAT project folder on the hard drive
(i.e. txtinout folder)}

\item{run_path}{Path to the SWAT project run folder}
}
\description{
Prepare run info for the simulation experiment
}
\keyword{internal}
