% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_handling.R
\name{load_demo}
\alias{load_demo}
\title{Loading SWAT demo data}
\usage{
load_demo(dataset, path = NULL, version = NULL, revision = NULL)
}
\arguments{
\item{dataset}{Character string to define data set to load. Valid inputs are
\code{dataset = c('project', 'observation', 'subbasin', 'river', 'hru')}.
\code{dataset = 'project'} loads a SWAT demo project in the defined
\code{path} location and returns the final project path as a text string in
R. The definition of the \code{version} is required.
\code{dataset = 'observation'} returns a \code{tibble} with discharge observation
data at the main outlet of the demo watershed.
The options \code{'subbasin'}, \code{'river'}, and \code{'hru'} return the paths
of the respective subbasin, river network and HRU shape files. See examples
how to load shapes unsing e.g. the \code{sf} package. These options require
a definition of the \code{version}.}

\item{path}{Character string that defines the path where copy the SWAT demo
project.}

\item{version}{Character string that defines the SWAT version. Options
are \code{version = c('2012', 'plus')}. This argument is required to
load SWAT projects and shape files.}

\item{revision}{Numeric value to define the model revision ofthe loaded SWAT
project. See \href{https://github.com/chrisschuerz/SWATdata}{SWATdata} for valid revision numbers.
When loading a SWAT project and leaving \code{revision = NULL}, then the
most recent model revision is loaded by default.}
}
\description{
With this function you can load data sets for SWAT2012 and SWAT+ demo projects.
The provided data includes demo projects, discharge observation data and spatial
data such as the subbasin boundaries, the river networks, and the HRUs of the
SWAT setups.
}
\section{Examples}{

To learn how to load the different demo data sets with \code{SWATplusR} see the
section \href{https://chrisschuerz.github.io/SWATplusR/articles/SWATplusR.html#loading-demos}{Loading demos}
on the package's \emph{Get Started} page.
}

