% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_parameter_read.R
\name{read_swat2012_files}
\alias{read_swat2012_files}
\title{Read the original swat parameter values from the parameter files in
project_path}
\usage{
read_swat2012_files(project_path, file_meta)
}
\arguments{
\item{project_path}{Path to the SWAT project folder on the hard drive
(i.e. txtinout folder)}

\item{file_meta}{Value supplied as file_meta to this internal helper.}
}
\description{
Read the original swat parameter values from the parameter files in
project_path
}
\keyword{internal}
