% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_run_swat.R
\name{run_swat2012}
\alias{run_swat2012}
\title{Run simulations for a SWAT2012 project}
\usage{
run_swat2012(
  project_path,
  output,
  parameter = NULL,
  start_date = NULL,
  end_date = NULL,
  output_interval = NULL,
  years_skip = NULL,
  rch_out_var = NULL,
  sub_out_var = NULL,
  hru_out_var = NULL,
  hru_out_nr = NULL,
  run_index = NULL,
  run_path = NULL,
  n_thread = NULL,
  save_file = NULL,
  save_path = NULL,
  return_output = TRUE,
  add_parameter = TRUE,
  add_date = TRUE,
  split_units = TRUE,
  run_in_project = FALSE,
  refresh = TRUE,
  keep_folder = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{project_path}{Path to the SWAT project folder on the hard drive
(i.e. txtinout folder).}

\item{output}{Output variables to extract and return for each SWAT model run.

Each output variable which should be returned is defined with \code{\link{define_output}}.
If multiple output variables should be returned for a model run, the output
definitions must be concatenated in a named list. See the examples in the
section below and on the help page of \code{\link{define_output}}.}

\item{parameter}{(optional) Model parameter changes to be implemented in the
simulation runs.

Parameter changes can be defined by providing a named vector with the
changes for a single parameter combination or a tibble where each row is
a parameter combinantion which should be used in a model simulation.

For examples of the implementation of parameter changes in model simulations
see the
\href{https://chrisschuerz.github.io/SWATplusR/articles/SWATrunR.html}{\code{SWATrunR}}
github page and the examples in corresponding section below.}

\item{start_date, end_date}{(optional) Start and end dates of the SWAT simulation.

Start and end dates have to be defined together. Default values are
\code{NULL}. In this case the start and end dates which are defined in file
'time.sim' of the SWAT+ project are used in the simulation runs.

The dates can be provided as character strings in any ymd format
(e.g. 'yyyy-mm-dd'), as numeric values (yyyymmdd) or in a \code{Date}
format.}

\item{output_interval}{(optional) Time interval in which the SWAT model
outputs are written.

If \code{NULL} (default) the value which is defined in the file.cio is used.
\code{output_interval} can be defined either with a character string where \code{'d'}
would result in daily output, and \code{'m'} in monthly, and \code{'y'} in yearly
outputs. Alternatively also the as SWAT2012 input values (0 for monthly,
1 for daily, 2 for yearly) are allowed.}

\item{years_skip}{(optional) Number of years to skip before printing outputs.

\code{years_skip} can be used to define the start year in which the first
simulation outputs should be written. Hence, \code{years_skip} is useful to
define a 'warm-up' period for the simulation runs for which the simulation
results are not considered.

\code{years_skip} is provided as an integer value which defines the number of
years to skip before writing outputs.

The default values of \code{years_skip} is \code{NULL}. In this case the projects'
default setting for output printing is used which is defined in the input
file file.cio'.}

\item{rch_out_var}{(optional) Numeric vector of maximum \code{length = 20} for
customized output of reach variables.

For output codes see the
\href{https://swat.tamu.edu/media/69308/ch03_input_cio.pdf}{SWAT I/O
Documentation} p.77ff.}

\item{sub_out_var}{(optional) Numeric vector of maximum \code{length = 15} for
customized output of subbasin variables.

For output codes see the
\href{https://swat.tamu.edu/media/69308/ch03_input_cio.pdf}{SWAT I/O
Documentation} p.78ff.}

\item{hru_out_var}{(optional) Numeric vector of maximum \code{length = 20} for
customized output of HRU variables.

For output codes see the
\href{https://swat.tamu.edu/media/69308/ch03_input_cio.pdf}{SWAT I/O
Documentation} p.79ff.}

\item{hru_out_nr}{(optional) Numeric vector of maximum \code{length = 20} for
providing the HRU numbers for which the HRU variables are written.

Optional if \code{hru_out_nr = 'all'}, HRU variables are written for all HRU
(caution, very large output files possible!)}

\item{run_index}{(optional) Vector of IDs of parameter combinations for which
simulations should be performed.

\code{run_index} is useful to only perform a subset of simulations for a defined
parameter set (e.g. when distributing simulations between several computers,
or when rerunning failed simulations for certain parameter combinations).

\code{run_index} is defined as a numeric vector where the values correspond to
the row IDs of the parameter input table. (e.g.\code{run_index = c(1:100,
  110, 115)}). The default value is \code{NULL}. Then the simulations for all
provided parameter combinations are performed.}

\item{run_path}{(optional) Path to a location where the simulations should be
performed.

\code{run_path} is useful if the project folder (in \code{project_path}) is in a
location where the simulations should not be performed (e.g. because of
limited space, or a virtual or slow drive). In such a case it can help to
perform the simulations on a different hard drive (e.g. local drive).

In the path which is defined with \code{run_path} the folder '.model_run' is
generated and the SWAT project is copied there into a number of sub-folders
'thread_\if{html}{\out{<i>}}', depending on the number of parallel runs which were defined.

With the default value of \code{run_path = NULL} the folder '.model_run' is
generated in the \code{project_path}.}

\item{n_thread}{(optional) Number parallel threads in which the simulations
are performed. Default \code{n_thread} is \code{NULL} and all simulations are run in
the folder '.model_run/thread_1' and on a single core. If \code{n_thread} is
defined with a number greater than 1 multiple thread folders are generated
and the simulations are performed in parallel on that number of cores.
\code{n_thread} is always limited by the number of cores of the used computer.}

\item{save_file}{(optional) Name of the folder where simulation runs should be
saved incrementally.

If a \code{save_file} is provided a folder with that name is generated and the
simulation results are incrementally saved in several SQLite data bases in
the folder. This option is useful when very long simulation experiments are
performed and computer/simulation crashes may result in the loss of days or
weeks of simulation runs, or when simulation outputs are returned which
would not fit into the computers RAM.

Default is \code{NULL} and simulations are not saved incrementally.}

\item{save_path}{(optional) Path where the \code{save_file} folder should be
generated.

In the default case \code{save_path = NULL} the \code{save_file} folder is generated
in the \code{project_path}. If a \code{save_path} is provided the \code{save_file} folder
is generated there if \code{save_file} in not \code{NULL}.}

\item{return_output}{(optional) Return outputs in \code{R}. If \code{TRUE} (default
value) then the defined simulated output variables are returned at the end
of all simulation runs. If \code{FALSE} no simulations are returned in \code{R}.

\strong{Caution:} Please only set \code{return_output = FALSE} if the simulations are
saved in a \code{save_file}. Otherwise the simulations are performed but no
simulations are returned.}

\item{add_parameter}{(optional)  Add the used parameter changes in the
simulation outputs. If \code{TRUE} (default value and recommended) the returned
list with the simulation outputs includes the list element \code{.$parameter}
which provides the definition of the parameters \code{.$parameter$definition}
and the values of the parameter changes \code{.$parameter$values}.}

\item{add_date}{(optional) Add a date vector to simulated time series. If
\code{TRUE} (default value) a date vector is added.}

\item{split_units}{(optional) Split simulated outputs of the same variable into
separate tables in the \code{.$simulation} list. If \code{TRUE} (default value) outputs
are split (e.g. for \code{define_output('rch', 'FLOW_OUT', 1:3)}) the
output tables \code{FLOW_OUT_1}, \code{FLOW_OUT_2}, and \code{FLOW_OUT_3} will be generated.
If \code{FALSE} in the same example a single table is generated with an integer
column \code{unit} with the ID values 1 to 3.}

\item{run_in_project}{Should a simulation be executed directly in the project
folder? If \code{FALSE} (default) the subfolder structure '.model_run/thread_*'
is generated and simulations are run in the thread folders. If \code{TRUE} the
simulation will be performed directly in the \code{project_path}. This is only
allowed for single simulation runs (No or a single parameter set provided).
\strong{Caution:} This option can overwrite the original model input files!}

\item{refresh}{(optional) Rewrite existing '.model_run' folder. If \code{TRUE}
(default value and recommended) always forces that .model_run' is newly
written when SWAT run ins started.}

\item{keep_folder}{(optional) Keep the thread folders in '.model_run' after
finishing the simulation runs. If \code{FALSE} (default value) the '.model_run'
folder is deleted after the last simulation. Keeping the thread folders
(\code{keep_folder = TRUE}) can be useful to manually check input and output
files in the thread folders after the simulation.}

\item{quiet}{(optional) Printing the progress of the simulations. If
\code{TRUE} (default value) the progress of the simulation runs is printed in the
console.}
}
\value{
Returns the simulation results as list with the following elements:
\itemize{
\item \code{.$parameter} is only available if parameter changes were implemented in
the simulation runs. \code{.$parameter} is a list with 2 elements:
\itemize{
\item \code{.$parameter$definition} is a tibble which shows how the parameter
changes were defined.
\item \code{.$parameter$values} is a tibble with the values of the defined
parameter changes. Each row is a parameter set. The names of the
columns correspond to the \code{par_name} which are defined in
\code{.$parameter$definition}.
}
\item \code{.$simulation} is a list with the defined and simulated output variables.
Each list element is one defined output variable. The name of each list
element is the name which was defined with \code{define_output()}.
\item \code{.$error_report} is only available if simulation runs failed. This element
is a tibble which summarizes the failed simulations and the triggered errors
of the SWAT executable which caused a simulation to fail.
\item \code{.$run_info} is a list with meta information on the simulation run. The
saved information is:
\itemize{
\item \code{.$run_info$simulation_log} is a tibble which logs simulation start and
end time_stamps, duration and paths of the project folder.
\item \code{.$run_info$simulation_period} is a tibble which provides dates, and skipped
years of the saved output variables.
\item \code{.$run_info$output_definition} is a tibble which summarizes the defined
output variables which were passed with the input argument \code{output} and
defined with \code{define_output()}.
}
}
}
\description{
\code{run_swat2012()} performs simulations for a SWAT2012 model setup that is located
in the folder \code{project_path} and returns simulated output variables which are
defined with \code{output} in a tidy format. Optionally, a set of parameter changes
can be defined with \code{parameter} which are then applied in the simulation runs.

Several settings such as the definition of the simulation period, incremental
saving of simulation results into SQLite data bases, or the use of multiple
cores for the simulation runs can be defined.
}
\section{Examples for output definition}{


Outputs are defined with the function \code{define_output()}. The function has the
three input arguments \code{file} which defines the output file into which the
variable of interest is written, \code{variable} which is the variable name as it is
defined in the output file (without the unit), and \code{unit} which is the spatial
unit for which the variable is extracted (e.g. channel ID, or HRU, etc.).

The following examples show definitions for output variables. In a simulation
run with \code{run_swat2012()} these definitions must be passed with the input
argument \code{output}.

A single output variable can be defined like in the following examples:

\if{html}{\out{<div class="sourceCode">}}\preformatted{# Define ET for all subbasins (e.g. if the setup has 125 subbasins)
# subbasin ET is in written into output.sub.
# The definition of the time interval in which the outputs are written are
# defined with the input argument output_interval (see arguments above)
define_output(file = 'sub',
              variable = 'ET', # Variable name in the file (capital letters!)
              unit = 1:125 # All subbasins 1 to 125
              )

# Define discharge for the channels 1 to 3 and 7
# Channel outputs are written into output.rch.
 define_output(file = 'rch',
               variable = 'FLOW_OUT',
               unit = c(1:3, 7) # To define the channel IDs of interest.
              )
}\if{html}{\out{</div>}}

More than one variable must be defined in a named list. The name of each
output definition is then assigned to the returned outputs together with the
unit ID if more than one ID was defined.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# Define the outputs for
# - ET for the HRUs 1 to 847
# - Channel discharge for the channels 1, 34, and 57
list(et_hru  = define_output('hru', 'ET', 1:847),
     cha_q   = define_output('rch', 'FLOW_OUT', c(1, 34, 57)))
}\if{html}{\out{</div>}}
}

\section{Examples for parameter definition}{


The definition of parameter changes follows a strict syntax in the parameter
names which have to be assigned to the change values. Therefore it is highly
recommended to read the corresponding section on the
\href{https://chrisschuerz.github.io/SWATrunR/articles/SWATrunR.html#changing-parameter-values}{\code{SWATrunR}}
github page. The following examples show very basic parameter definitions and
do not cover specific naming of parameters, or parameter conditions which are
topics covered on the github page.

In a nutshell the name definition of a parameter requires at least the name
of the parameter how it is defined in the calibration.cal input file, the
file suffix, and the type of change, as shown in this example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{# Change CN2 by -5\% and ALPHA_BF by an absolute value of 0.35
par_set <- c("CN2.mgt | change = relchg" = - 0.05,
"            "ALPHA_BF.gw | change = absval" = 0.35)
}\if{html}{\out{</div>}}

As shown in the example above, single parameter changes can be passed as a
named vector. Multiple parameter combinations must be passed in a tibble
(caution here, base R data.frames might not be able to handle all special
characters in the parameter names). Below an example for a more comprehensive
parameter set with uniform randomly sampled parameter changes:

\if{html}{\out{<div class="sourceCode">}}\preformatted{library(tibble)

# Draw 100 samples for each parameter
n <- 100

# Sample each parameter change uniformly n times in their upper/lower bounds
par_set <- tibble('CN2.mgt | change = abschg' = runif(n,-15,10),
                  'CANMX.hru | change = absval |
                  'EPCO.bsn | change = absval' = runif(n,0.1,1),
                  'ESCO.bsn | change = absval' = runif(n,0.1,1),
                  "ALPHA_BF.gw | change = absval" = runif(n,0.01,0.8))
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# Install the SWATdata R package which provides a SWAT+ demo project
if(!'SWATdata' \%in\% installed.packages()) {
  remotes::install_github('chrisschuerz/SWATdata')
}

# Use a temporary dir for the demo project
tmp_dir <- tempfile("swatrunr-demo-")
dir.create(tmp_dir)

# Load a SWAT2012 demo project
proj_path <- load_demo(dataset = 'project',
                       path = tmp_dir,
                       version = '2012')

# Perform simulations for the demo project and return
# daily simulated discharge for channel 1
q_sim <- run_swat2012(project_path = proj_path,
                      output = define_output(file = 'rch',
                                             variable = 'FLOW_OUT',
                                             unit = 1))

# The simulated time series in a tibble format
q_sim$simulation$FLOW_OUT

# Plot the simulated time series
plot(q_sim$simulation$FLOW_OUT, type = 'l')

# Print the meta information for the simulation run
q_sim$run_info

# Delete demo project folder
unlink(tmp_dir, recursive = TRUE, force = TRUE)
}

}
