% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_parameter_read.R
\name{split_line}
\alias{split_line}
\title{Split one line in a parameter file into the individual values of the table}
\usage{
split_line(chr, start, end)
}
\arguments{
\item{chr}{Value supplied as chr to this internal helper.}

\item{start}{Index vector indicating the start values of a value}

\item{end}{Index vector indicating the end values of a value}
}
\description{
Split one line in a parameter file into the individual values of the table
}
\keyword{internal}
